airflow/requirements.txt
    boto3,
    numpy,
    pandas,
    python-dotenv,
    scikit-learn,
    